Wove a gentler index
