+++ docs/genesis/seed/transmission.html
@@ -4,6 +4,7 @@
   <meta charset='UTF-8'>
   <title>Transmission</title>
   <link rel='stylesheet' href='../../style.css'>
+  <script src='../../main.js' defer></script>
 </head>
 <body>
 <article>
