+++ docs/index.html
@@ -4,15 +4,36 @@
   <meta charset='UTF-8'>
   <title>LichenLoom</title>
   <link rel='stylesheet' href='style.css'>
+  <script src='main.js' defer></script>
 </head>
 <body>
 <article>
 <h1>LichenLoom</h1>
-<ul>
-<li><a href='genesis/spore/micro_story.html'>Micro Story</a></li>
+<ul class=''>
+<li class='branch'>genesis
+<ul class='hidden'>
+<li class='branch'>seed
+<ul class='hidden'>
 <li><a href='genesis/seed/opening.html'>Opening</a></li>
 <li><a href='genesis/seed/transmission.html'>Transmission</a></li>
+</ul>
+</li>
+<li class='branch'>spore
+<ul class='hidden'>
+<li><a href='genesis/spore/micro_story.html'>Micro Story</a></li>
+</ul>
+</li>
+</ul>
+</li>
+<li class='branch'>odyssey
+<ul class='hidden'>
+<li class='branch'>whispers
+<ul class='hidden'>
 <li><a href='odyssey/whispers/embark.html'>Embark</a></li>
+</ul>
+</li>
+</ul>
+</li>
 </ul>
 </article>
 </body>
