+++ docs/style.css
@@ -1,1 +1,1 @@
-body{font-family:sans-serif;background:#fdfdfd;color:#222;margin:2rem;}article{max-width:60ch;margin:auto;}h1,h2,h3{font-family:serif;color:#00524f;}
+body{font-family:'Helvetica Neue',sans-serif;background:#f0f0f0;color:#222;margin:2rem;line-height:1.6;}article{max-width:65ch;margin:auto;}h1,h2,h3{font-family:'Georgia',serif;color:#003b30;}a{color:#00584d;text-decoration:none;}a:hover{text-decoration:underline;}ul{list-style:none;padding-left:1rem;margin:0;}li{margin:.25rem 0;} .branch>ul{margin-left:1rem;border-left:1px dotted #ccc;padding-left:.5rem;}span.toggle{cursor:pointer;margin-right:.25rem;color:#003b30;transition:transform .2s;display:inline-block;}li.open>span.toggle{transform:rotate(90deg);}ul.hidden{display:none;}
